requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
yt-dlp
python-dotenv
slowapi